first approach of Graph object generation. Link to system missing
I don't know how I want to to do the connection to the system right now.
The current approach would need  two model classes, two forms, two views
and all for whether it is a file or a quick system. thats not nice.
Also don't know how and what to use as a foreign key.